Restructure some components & minor css changes
