--- src/components/Content/About/About.css
@@ -1,4 +1,0 @@
-.about-content {
-    min-height: 100vh;
-    padding: 4rem;
-}
