--- src/components/Content/Contact/Contact.css
@@ -1,3 +1,0 @@
-.contact-content {
-    padding: 4rem;
-}
