+++ src/components/Content/Contact/Contact.js
@@ -2,7 +2,7 @@ import React from 'react';
 import './Contact.css';
 
 const Contact = () => (
-    <div id='contact' className='contact-content'>
+    <div id='contact' className='content-section'>
         <h1>Contact</h1>
         <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Nihil eius repellat quis maxime consequuntur natus tempore esse fugiat ex ratione. Ducimus ullam amet veritatis deserunt et totam est, ratione non culpa a repellendus nihil soluta tempora, numquam quis. Dignissimos officia corrupti eos illo, veritatis ullam dolores nesciunt ipsam temporibus consequatur quia culpa itaque distinctio aliquid reiciendis modi vel deleniti beatae quis, et facere non pariatur. Magni incidunt obcaecati sequi quam veniam dolor placeat molestiae fuga eaque dignissimos, consequatur quas assumenda inventore repellat at voluptatem blanditiis, cumque vitae beatae, rem amet ut expedita eos. Mollitia possimus doloremque quo iusto quaerat ea adipisci fugit vel corrupti recusandae sunt sapiente, reprehenderit excepturi voluptas non laboriosam obcaecati amet. Rem totam illum, tenetur ducimus veritatis ea fuga quisquam blanditiis pariatur molestiae eligendi cupiditate quos. Nobis aliquam eos facilis quaerat, maiores, ut molestiae ab repellendus praesentium quibusdam labore tempore eius! Culpa, fugit officia assumenda, quo tempora fugiat laudantium in cumque nesciunt ea expedita quas voluptatum sapiente blanditiis neque? Reiciendis modi minima beatae veniam, veritatis obcaecati dolor soluta et. Quae facere doloremque libero suscipit consequuntur distinctio repellendus inventore eos dolor ea soluta fugiat provident, vel iusto est molestias consectetur animi aspernatur adipisci. Dignissimos consequatur assumenda recusandae delectus.</p>
     </div>
