+++ src/components/Content/Frontpage/About/OurTeam/MemberCard/MemberCard.css
