+++ src/components/Content/Frontpage/Frontpage.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import Header from '../Header/Header';
-import About from '../About/About';
-import Games from '../Games/Games';
+import Header from './Header/Header';
+import About from './About/About';
+import Games from './Games/Games';
 
 const Frontpage = () => (
     <div>
