+++ src/components/Content/Frontpage/Games/Games.js
@@ -1,8 +1,7 @@
 import React from 'react';
-import './Games.css';
 
 const Games = () => (
-    <div id='games' className='games-content'>
+    <div id='games' className='content-section'>
         <h1>Games</h1>
         <p>Lorem ipsum, dolor sit amet consectetur adipisicing elit. Ipsa voluptas porro maiores natus dicta atque beatae, eius, praesentium, illum error cumque repellendus sapiente nisi quis nam veniam! Libero, soluta qui nesciunt delectus neque, quo fugit, recusandae iste voluptate perferendis fuga numquam incidunt eum accusamus maiores distinctio quidem accusantium tempora non aperiam aut. Cum impedit dolorem optio quis consequuntur ut, iusto repudiandae rerum eligendi quaerat assumenda! Voluptatibus cupiditate amet necessitatibus totam? Perspiciatis, voluptates! Voluptatem voluptas enim itaque aliquid quos, nulla perspiciatis sunt fugiat corporis. Reiciendis amet doloribus velit atque fugiat vero voluptate necessitatibus corrupti accusantium rerum facilis quae in, placeat eum?</p>
     </div>
