--- src/components/Content/Games/Games.css
@@ -1,4 +1,0 @@
-.games-content {
-    min-height: 100vh;
-    padding: 4rem;
-}
